# docs/43-api-contract.openapi.yaml
openapi: 3.1.0
info:
  title: Shikhi API
  version: 0.2.0-draft
  description: >
    REST API contract for Shikhi (শিখি) — a Bengali→English learning platform.
    Single source of truth for the Spring Boot backend, the React SPA (ADR-0002/0007),
    and the native Android client (ADR-0012).
    Status: DRAFT — reviewed at Gate B. Endpoints marked `x-phase: post-pilot` are
    designed-for but may ship after the curated pilot (D6). AI grading is NOT in this
    contract (D4) — it is additive behind the grading seam and does not change these
    shapes (ADR-0006, NFR-M6).
servers:
  - url: https://api.shikhi.example/v1
    description: Production (placeholder; region/provider per ADR-0008)
  - url: http://localhost:8080/v1
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /practice/sessions/{sessionId}/answers:
    post:
      tags: [practice]
      summary: Submit an answer to a generated exercise (idempotent; verdict + updated stats)
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SubmitAnswerRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AnswerResult' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /practice/sessions/{sessionId}/answers/batch:
    post:
      tags: [practice]
      summary: Submit all buffered answers at once; server re-grades authoritatively (idempotent per key), returns final stats + per-exercise verdicts (E12, ADR-0013)
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PracticeAnswerBatch' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PracticeBatchResult' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /practice/sessions/{sessionId}/rounds:
    post:
      tags: [practice]
      summary: Generate the next round of exercises ("Keep going")
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PracticeRound' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /practice/sessions/{sessionId}/complete:
    post:
      tags: [practice]
      summary: Finish a practice session (idempotent; totals + level-up eligibility)
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/IdempotentRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PracticeResult' } } } }

  # ---------------- REVIEW (spaced repetition) ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Conflict: { description: Conflict, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      required: [code, message, correlationId]
      properties:
        code: { type: string, example: VALIDATION_FAILED }
        message: { type: string, description: Localized (BN/EN) human-readable message }
        details: { type: object, additionalProperties: true }
        correlationId: { type: string }
    Bilingual:
      type: object
      properties:
        en: { type: string }
        bn: { type: string }
...
    IdempotentRequest:
      type: object
      required: [idempotencyKey]
      properties: { idempotencyKey: { type: string } }
    SubmitAnswerRequest:
      allOf:
        - $ref: '#/components/schemas/IdempotentRequest'
        - type: object
          required: [exerciseId, answer]
          properties:
            exerciseId: { type: string, format: uuid }
            answer:
              type: object
              description: Shape depends on exercise type (e.g., {selectedOptionId} or {text} or {tokenOrder})
              additionalProperties: true
    Verdict:
      type: object
      properties:
        correct: { type: boolean }
        feedback: { $ref: '#/components/schemas/Bilingual' }
        matchedPatternCode: { type: string, nullable: true }
        source: { type: string, enum: [RULE, AI], default: RULE }
    AnswerResult:
      type: object
      properties:
        verdict: { $ref: '#/components/schemas/Verdict' }
        stats: { $ref: '#/components/schemas/Stats' }
...
    Stats:
      type: object
      properties:
        xp: { type: integer }
        rank: { type: integer }
        currentStreak: { type: integer }
        longestStreak: { type: integer }
        hearts: { type: integer }
        dailyGoal: { type: integer }
        cefrLevel: { type: string, enum: [A1, A2, B1, B2, C1], description: 'Learner CEFR band (E12; C1 admitted in V21, 2026-07-08)' }
        accuracyByPattern:
          type: object
          additionalProperties: { type: number }
...
    PracticeExercise:
      type: object
      description: >
        A generated vocabulary exercise (E12). Type-specific presentation fields live under
        `config` (never any correctness flag). Unlike graded lessons, practice also serves a
        `solution` so the web client can grade instantly and locally (ADR-0013); the server
        still re-grades authoritatively on submit.
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [WORD_MEANING, MEANING_WORD, SENTENCE_GAP, SENTENCE_BUILD, TYPE_WORD] }
        ordinal: { type: integer }
        prompt: { $ref: '#/components/schemas/Bilingual' }
        config: { type: object, additionalProperties: true, description: 'e.g., options[], tokens[] (no correctness flags)' }
        solution: { $ref: '#/components/schemas/PracticeSolution' }
    PracticeSolution:
      type: object
      description: >
        The correct answer for a practice exercise (ADR-0013). Served for practice only; graded
        lesson sessions keep a server-only answer key.
      properties:
        correctOptionId: { type: string, nullable: true, description: 'Correct option id for MCQ types (WORD_MEANING, MEANING_WORD, SENTENCE_GAP)' }
        accepted: { type: array, items: { type: string }, nullable: true, description: 'Accepted normalized answers for SENTENCE_BUILD / TYPE_WORD' }
        reveal: { type: string, description: 'Correct-answer text shown in the wrong-answer reveal' }
    PracticeRound:
      type: object
      properties:
        sessionId: { type: string, format: uuid }
        round: { type: integer }
        cefrLevel: { type: string, enum: [A1, A2, B1, B2], description: 'Level pinned at session start' }
        exercises: { type: array, items: { $ref: '#/components/schemas/PracticeExercise' } }
        levelUpEligible: { type: boolean, description: 'True when mastery suggests moving up a band (US-12.7)' }
    PracticeResult:
      type: object
      properties:
        correctCount: { type: integer }
        totalCount: { type: integer }
        roundsPlayed: { type: integer }
        xpEarned: { type: integer }
        levelUpEligible: { type: boolean }
        stats: { $ref: '#/components/schemas/Stats' }
    PracticeAnswerBatch:
      type: object
      required: [answers]
      description: All answers buffered client-side for one flush (E12, ADR-0013).
      properties:
        answers:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SubmitAnswerRequest' }
    PracticeBatchResult:
      type: object
      properties:
        stats: { $ref: '#/components/schemas/Stats' }
        verdicts:
          type: array
          items:
            type: object
            properties:
              exerciseId: { type: string, format: uuid }
              correct: { type: boolean }
...